requirements.txt
fastapi~=0.118.0
uvicorn[standard]~=0.37.0
pydantic~=2.12.3
python-dotenv~=1.1.1
google-generativeai
pymongo~=4.15.3
weaviate-client~=4.17.0
python-jose[cryptography]~=3.5.0
passlib[bcrypt]
python-multipart~=0.0.20
bson~=0.5.10
pydantic_core~=2.41.4
protobuf~=5.29.5
h11~=0.16.0
pip~=25.0.1
attrs~=25.3.0
distro~=1.9.0
cryptography~=46.0.2
pillow~=11.3.0
Jinja2~=3.1.6
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.75.1
typing_extensions~=4.14.1
requests~=2.32.5
rich~=14.2.0
colorama~=0.4.6
PyYAML~=6.0.3
sniffio~=1.3.1
proto-plus~=1.26.1
click~=8.2.1
aiohttp~=3.12.15
pyasn1_modules~=0.4.2
cachetools~=6.2.0
certifi~=2025.10.5
urllib3~=2.3.0
packaging~=24.2
cffi~=2.0.0
google-api-python-client~=2.184.0
grpcio-status~=1.71.2
httplib2~=0.31.0
tqdm~=4.67.1
starlette~=0.48.0
websockets~=15.0.1
anyio~=4.11.0
orjson~=3.11.3
httptools~=0.6.4
watchfiles~=1.1.0
pyparsing~=3.2.5
annotated-types~=0.7.0
idna~=3.10
charset-normalizer~=3.4.3
itsdangerous~=2.2.0
httpx~=0.28.1
Werkzeug~=3.1.3
uritemplate~=4.2.0
google-auth-httplib2~=0.2.0